+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - ADD: Search and paging in allergens view
 - ADD: Include full list of alergens in generated pdf
 - ADD: Allergens page improvements
+- ADD: Upload ingredients to PV Online
 - FIX: Not showing allergens correctly in label
 - FIX: Message position in Generate Finished Product page
 - FIX: Various bug fixes